chore(RuleEngine): split ruleengine component in different components
